StreamUtils: move to `o.b.b.internal` from `o.b.b.utils`
